Extract filters to "eleventy.filters.mjs"
